Rename Utils/getUser.ts -> Utils/user.ts
