+++ src/Components/Artwork/GridItem.tsx
@@ -7,7 +7,7 @@ import React from "react"
 import { createFragmentContainer, graphql } from "react-relay"
 import { data as sd } from "sharify"
 import styled from "styled-components"
-import { userIsAdmin } from "Utils/getUser"
+import { userIsAdmin } from "Utils/user"
 import Badge from "./Badge"
 import Metadata from "./Metadata"
 import SaveButton from "./Save"
